Cloud Run Functions Sample Apps (#9)
## What type of PR is this? (check all applicable)

- [x] Feature / Major Change / Refactor / Optimization
- [ ] Bug Fix
- [ ] Documentation Update

## Description

- Create sample apps for .NET, Go, Java, Node, Python, and Ruby

Caveats:
- In Java, the `dd-java-agent.jar` needs to be deployed with the
function. So it's installed via `curl` and committed in the repo. This
should be fine since it's only ~30mb. This seems to be the cleanest
solution -- alternatively, the user could install the tracer via
Gradle/Maven and use a plugin to copy the jar over to the cwd, but the
diff for that approach is pretty huge, and I think for documentation
purposes, telling users to `curl` the tracer agent is simplest.
- PHP is not supported since we don't have a Dockerfile, and there's no
package manager like NPM or PyPi to install the tracer from.
Additionally, the tracer is installed in `/usr/local/lib/php`; it's not
a simple file like Java we can just add to our deployment.
- For PHP, technically the user could (1) spin up a Linux container via
Docker, (2) run the PHP setup script which installs the tracer in system
directories, (3) move the files from `/usr/local/lib/php/extensions/`
and `/usr/local/etc/php/conf.d/`, and (4) commit the files and deploy
the files with the function. [See [Tracing PHP
Applications](https://docs.datadoghq.com/tracing/trace_collection/automatic_instrumentation/dd_libraries/php)
for more info]. But this is getting overly complex, and there's hardly
any demand for PHP. We can revisit adding documentation for PHP in the
future if people start requesting it.

## Related tickets & Documents

<!--
For new features or major changes, we follow a documented RFC process
(Check our CONTRIBUTING guidelines).
We cannot accept new features or major changes without a linked approved
RFC.

For pull requests that relate or close an issue, please include them
below.
-->

- Approved RFC (for feature/major changes) #
- Related Issue #
- Closes #

## How to reproduce and testing

<!--
For bug fixes, please describe how you reproduced the issue, what
environments this was tested on (architecture, OS, etc.), and how the PR
solves the issue.
-->